Restructure download handler tests.
